MaterialProviders_stop/ComAmazonawsDynamodb/runtimes/python/pyproject.toml
[tool.poetry]
name = "amazon-cryptography-internal-dafny-shim-ddb"
version = "0.1.0"
description = ""
authors = ["AWS <aws-arg-dafny@amazon.com>"]
packages = [
    { include = "com_amazonaws_dynamodb", from = "src" }
]
# Include all of the following .gitignored files in package distributions,
# even though it is not included in version control
include = ["**/smithygenerated/*.py", "**/dafnygenerated/*.py"]

[tool.poetry.dependencies]
python = "^3.11.0"
boto3 = "^1.28.38"
amazon-cryptography-internal-dafny-shim-ddb-externs = { path = "src/com_amazonaws_dynamodb/extern", develop = false}
# TODO: Depend on PyPi once Smithy-Python publishes their Python package there
smithy-python = { path = "../../../../../submodules/smithy-python/python-packages/smithy-python", develop = false}
standard-library = { path = "../../../StandardLibrary/runtimes/python", develop = false}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
amazon-cryptography-internal-dafny-shim-ddb-test-internaldafny = { path = "test/internaldafny", develop = false}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
